ISFNet/requirements.txt
antlr4-python3-runtime==4.9.3
certifi==2025.10.5
charset-normalizer==3.4.4
connected-components-3d==3.26.1
contourpy==1.3.2
cycler==0.12.1
einops==0.8.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.9.0
hf-xet==1.1.10
huggingface-hub==0.35.3
hydra-core==1.3.2
idna==3.11
imageio==2.37.0
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
MarkupSafe==3.0.3
matplotlib==3.10.7
MedPy==0.5.2
monai==1.5.1
mpmath==1.3.0
networkx==3.4.2
nibabel==5.3.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-python-headless==4.11.0.86
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
pysodmetrics==1.5.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
safetensors==0.6.2
scikit-image==0.24.0
scipy==1.15.3
simpleitk==2.5.2
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
timm==1.0.20
torch==2.4.1
torchaudio==2.4.1
torchvision==0.19.1
tqdm==4.65.2
triton==3.0.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
